requirements.txt
fastapi[all]
uvicorn
google-analytics-data
google-auth
openai
loguru
dotenv